added EvaporationModel class
